Initial Commit
I created this demo for a question on Stack Overflow.  The idea was to
create a menu system where you had a large number of selectable items
spread over several pages.  Like a level selector.  The MainScene
creates a grid of NxM elements per page, with S pages total, and places
next/back buttons for each.  All the elements are actually created at
once, but are off camera.  The solution slides the items in/out as the
user presses the next/back buttons.